fix: deduplicate id, return updated data

+++ packages/adapter-mongodb/src/index.ts
@@ -47,15 +47,13 @@ export const format = {
   /** Takes a plain old JavaScript object and turns it into a mongoDB object */
   to<T = Record<string, unknown>>(object: Record<string, any>) {
     const newObject: Record<string, unknown> = {
-      _id: _id(object.id),
+      _id: new ObjectId(object.id),
     }
     for (const key in object) {
       const value = object[key]
-      if (key === "userId") {
-        newObject[key] = _id(value)
-      } else {
-        newObject[key] = value
-      }
+      if (key === "userId") newObject[key] = _id(value)
+      else if (key === "id") continue
+      else newObject[key] = value
     }
     return newObject as T
   },
@@ -111,17 +109,20 @@ export function MongoDBAdapter(
       return from<AdapterUser>(user)
     },
     async updateUser(data) {
-      const { value: user } = await (
+      const { _id, ...user } = to<AdapterUser>(data)
+
+      const result = await (
         await db
-      ).U.findOneAndUpdate({ _id: _id(data.id) }, { $set: data })
-      return from<AdapterUser>(user!)
+      ).U.findOneAndUpdate({ _id }, { $set: user }, { returnDocument: "after" })
+
+      return from<AdapterUser>(result.value!)
     },
     async deleteUser(id) {
       const userId = _id(id)
       const m = await db
       await Promise.all([
         m.A.deleteMany({ userId }),
-        m.S.deleteMany({ userId }),
+        m.S.deleteMany({ userId: userId as any }),
         m.U.deleteOne({ _id: userId }),
       ])
     },
@@ -156,7 +157,11 @@ export function MongoDBAdapter(
     async updateSession(data) {
       const { value: session } = await (
         await db
-      ).S.findOneAndUpdate({ sessionToken: data.sessionToken }, { $set: data })
+      ).S.findOneAndUpdate(
+        { sessionToken: data.sessionToken },
+        { $set: data },
+        { returnDocument: "after" }
+      )
       return from<AdapterSession>(session!)
     },
     async deleteSession(sessionToken) {
